test_platform/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
streamlit>=1.20.0
pytest>=7.0.0
httpx>=0.27.0
requests>=2.31.0
tiktoken>=0.12.0
PyMuPDF>=1.23.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
openpyxl>=3.1.2
XlsxWriter>=3.1.9
xmind>=1.2.0
